feat: Implemented basic player scene
Introduces a basic player scene (player.tscn), associated script (player.gd), and art assets (aesprite and png files PlayerPlaceholder and PlayerPlaceholderDead).

In the script, the player can move with wasd keys and dash with space. There is also a basic take damage function, and when the player dies the sprite changes from alive to dead.