feat: Unibot is (theoretically) done
I have given up on Unibot moving in a circle around the player. Instead it moves vertically or horizontally (chosen randomly) in  a straight line, initially moving towards the player and reversing direction on wall collision. While moving it (hopefully) shoots bullets perpendicular to where the bot is moving towards the player.

Known issues: Bullets do not move in the correct direction when shot, which is a known bug with bullets. When that is fixed, I doubt the bullets shot by Unibot will work as expected because its kinda hard to test them right now.