feat: evil snake exists
I am working on making a snake enemy so we have something that shoots wavy bullets. The evil snake is set up to be worked on, but when creating bullets errors with "Invalid access to property or key 'bullet_sprite' on a base object of type 'Nil'." and I don't know why because I copied and pasted the code for shooting from enemy_base. Oh well.