feat: some work on evil snake
I did some work on making evil snake move and attack, but I have to do something else now so pushing what I did. My plan is to do a similar thing to Unibot, where it does nothing for a short time then attacks for a short time and swaps between those states. When evil snake is attacking I am planning on making it slowly move towards the player and shoot bullets in a wavy pattern, with the bullets moving in an alternating pattern of sin and -sin waves.