feat: Evil snake done
Evil snake is done. It will wait for one second, then moves slowly towards the player and shooting wavy bullets at them every 0.75 seconds for 2.5 seconds. It is done for the playtest, but I may try to change the bullets to move a little slower when curving and faster when moving straight